rename EntryDialect file
