+++ packages/site/src/routes/[dictionaryId]/entry/EntryDialect.svelte
